move dynamic_extraction into a separate file
